calculadora v2 agora com números decimais

+++ calculadorasimples.c
@@ -2,57 +2,57 @@
 
 int main(void)
 {
-    int n1;
-    int n2;
-    int soma, sub, mult, div;
-    int op;
+    double n1;
+    double n2;
+    double soma, sub, mult, div;
+    double op;
 
     printf("Seja bem-vindo. \nEssa é uma calculadora simples em linguagem C. \nVocê digita dois números inteiros e escolhe a operação que deseja realizar.\n");
     // printf();
     printf("Digite o primeiro número: ");
-    scanf("%d", &n1);
+    scanf("%lf", &n1);
     printf("Digite o segundo número: ");
-    scanf("%d", &n2);
+    scanf("%lf", &n2);
 
     printf("Qual operação quer realizar? \n1 - soma \n2 - subtração \n3 - multiplicação \n4 - divisão \n:");
-    scanf("%d", &op);
+    scanf("%lf", &op);
 
     if (op == 1)
     {
         soma = n1 + n2;
-        printf("A soma entre %d e %d é: %d", n1, n2, soma);
+        printf("A soma entre %lf e %lf é: %lf", n1, n2, soma);
     }
     else if (op == 2)
     {
         if (n1 > n2)
         {
             sub = n1 - n2;
-            printf("A subtração entre %d e %d é: %d", n1, n2, sub);
+            printf("A subtração entre %lf e %lf é: %lf", n1, n2, sub);
         }
         else if (n1 < n2)
         {
             sub = n2 - n1;
-            printf("A subtração entre %d e %d é: %d", n2, n1, sub);
+            printf("A subtração entre %lf e %lf é: %lf", n2, n1, sub);
             
         }
     }
     else if (op == 3)
     {
         mult = n1 * n2;
-        printf("A multiplicação entre %d e %d é: %d", n1, n2, mult);
+        printf("A multiplicação entre %lf e %lf é: %lf", n1, n2, mult);
     }
     else if (op == 4)
     {
         if (n1 > n2)
         {
             div = n1 / n2;
-            printf("A divisão entre %d e %d é: %d", n1, n2, div);
+            printf("A divisão entre %lf e %lf é: %lf", n1, n2, div);
 
         }
         else if (n1 < n2)
         {
             div = n2 / n1;
-            printf("A divisão entre %d e %d é: %d", n2, n1, div);
+            printf("A divisão entre %lf e %lf é: %lf", n2, n1, div);
 
         }
     }
